% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_optional_info.R
\name{add_optional_info}
\alias{add_optional_info}
\title{add columns to process tibble}
\usage{
add_optional_info(process, add = "all")
}
\arguments{
\item{process}{Process tibble to update}

\item{add}{Vector of info to add including RACI, Note and Branch}
}
\value{
process Updated process
}
\description{
Add all optional columns to process tibble
}
\examples{
p <- create_process(steps = 5, sipoc_template = TRUE)
p <- add_optional_info(p, add = "RACI")
p
p <- add_optional_info(p, add = "Level")
p
p <- add_optional_info(p)
p

}
