% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_process.R
\name{clean_process}
\alias{clean_process}
\title{Clean a process}
\usage{
clean_process(process)
}
\description{
Make sure all columns are present
Substitute out risky characters for their character code
}
