% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{colour_palette}
\alias{colour_palette}
\title{colour_palette}
\usage{
colour_palette(n = 5)
}
\arguments{
\item{n}{number of colours requested}
}
\value{
colours vector of colours in hex-format
}
\description{
Return a colour palette for process steps
}
