% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_process.R
\name{create_process}
\alias{create_process}
\title{Create a skeleton process tibble to use in process mapping}
\usage{
create_process(steps, levels = 2, sipoc_template = FALSE, add = NA)
}
\arguments{
\item{steps}{Sum of number of process steps across all levels}

\item{sipoc_template}{Include template information in SIPOC fields}

\item{add}{Vector of info to add including RACI, Note and Branch}

\item{max_level}{Process sub-level depth}
}
\value{
process with standard columns, as a tibble
}
\description{
This function makes it easier to get started with creating a process map by
creating a template data frame that can be saved and edited
}
\examples{
p <- create_process(steps = 5, sipoc_template = TRUE, add = "RACI")
p
p <- create_process(steps = 5, sipoc_template = TRUE, add = "RACI")
p
}
