% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{expand_palette}
\alias{expand_palette}
\title{expand_palette
Expand an vector of colours to be a specified length by repeating}
\usage{
expand_palette(palette, n = 5)
}
\arguments{
\item{palette}{vector of colours to expand}

\item{n}{number of colours requested}
}
\value{
colours vector of colours in hex-format
}
\description{
expand_palette
Expand an vector of colours to be a specified length by repeating
}
