% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_optional_info.R
\name{explain_optional_info}
\alias{explain_optional_info}
\title{explain meaning of the optional columns}
\usage{
explain_optional_info(info = "all")
}
\arguments{
\item{info}{Vector of info to explain including RACI, Note and Branch}
}
\value{
df Dataframe with Info and Description
}
\description{
Explain all optional columns that can be added to a process tibble
}
\examples{
df <- explain_optional_info(info = "RACI")
df
df <- explain_optional_info(info = "all")
df

}
