% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_text.R
\name{fmt_text}
\alias{fmt_text}
\title{fmt_text wrap text to specified width using defined character separator}
\usage{
fmt_text(text, wrap, sep = "&#92;n")
}
\arguments{
\item{text}{array containing text to wrap}

\item{wrap}{maximum width}

\item{sep}{separator string}
}
\value{
fmted array of wrapped text
}
\description{
utility function to wrap text
}
