% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_loopback_edges.R
\name{get_loopback_edges}
\alias{get_loopback_edges}
\title{get_loopback_edges generate edge dataframe for any loops}
\usage{
get_loopback_edges(process, edge_colour)
}
\arguments{
\item{process}{Process in standard format}
}
\value{
edge dataframe or NULL
}
\description{
Use To column to determine any additional edges to be created
called internally after clean_process and select_subprocess
To column includes destination in IDdotted format
}
