% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_navigation_map.R
\name{get_navigation_map}
\alias{get_navigation_map}
\title{Get named list of navigation nodes for a graph}
\usage{
get_navigation_map(graph)
}
\arguments{
\item{graph}{dgr graph}
}
\value{
named list
}
\description{
Get a list of all subprocesses as dataframe
}
\examples{
names <- get_navigation_map(graph)

}
