% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_sipoc.R
\name{graph_sipoc}
\alias{graph_sipoc}
\title{Creates a SIPOC graph from a process}
\usage{
graph_sipoc(
  process,
  xscale = 1.8,
  yscale = 0.9,
  sipoc_label_location = 0.25,
  step_textwrap = 22,
  title_fontsize = 14,
  edge_colour = base_box_palette,
  fill_colours = base_colour_palette,
  box_colours = NULL,
  font_colours = base_text_palette,
  fontname = "Calibri,Arial,Helvetica",
  title_fontname = "Calisto MT,Times New Roman",
  width = 1.35
)
}
\arguments{
\item{process}{A process structure}

\item{xscale}{Scale factor in x-dimension}

\item{yscale}{Scale factor in y-dimension}

\item{sipoc_label_location}{Where to plot the SIPOC legend box}

\item{step_textwrap}{Specify longest line before wrapping}

\item{title_fontsize}{Size of the title font in points (NA means no title)}

\item{edge_colour}{Colour of edeges/arrows}

\item{fill_colours}{Five (or one) colours for SIPOC boxes}

\item{box_colours}{Five (or one) colours for SIPOC boxes (default match fill colours)}

\item{font_colours}{Five (or one) colours for text in SIPOC boxes}

\item{fontname}{Name of font to use}

\item{title_fontname}{Name of font for the title}

\item{width}{Standard box width}
}
\description{
This function is user-callable and requires DiagrammeR. It takes a process data frame, including
Inputs, Outputs, Supplier and Customers and the L2 process and creates the nodes, egdes and a graph
as a SIPOC form in AZ colours from this input
}
\examples{
p <- create_process(steps = 6, sipoc_template = TRUE)
g <- graph_sipoc(p)
DiagrammeR::render_graph(g)

}
