% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_slm.R
\name{graph_slm}
\alias{graph_slm}
\title{Graph swimlane map}
\usage{
graph_slm(
  process,
  level = 1,
  parent = "",
  depth = NA,
  numbering = FALSE,
  terminals = TRUE,
  xscale = 1.5,
  yscale = 1.5,
  step_textwrap = 18,
  box_colour = "auto",
  box_style = "rounded",
  arrow_colour = base_arrow_palette,
  arrow_style = "solid",
  colour_palette = base_colour_palette,
  lane_colour = base_text_palette[1],
  title_fontsize = 14,
  fontname = "Calibri,Arial,Helvetica",
  title_fontname = "Calisto MT,Times New Roman",
  ctoffset = 0.3
)
}
\arguments{
\item{process}{Process to graph as a swimlane map}

\item{level}{The process level to plot (2-5)}

\item{parent}{The level of the process to plot}

\item{depth}{The deepest level to plot}

\item{numbering}{Number the process steps}

\item{terminals}{Start and End circles be shown}

\item{xscale}{Scale factor for x-dimension}

\item{yscale}{Scale factor for y-dimension}

\item{step_textwrap}{String length for wrapping text in process boxes (nodes)}

\item{box_colour}{Color for the edge of the process step box}

\item{box_style}{Box style for use by DiagrammerR::create_node_df
(rounded, solid, dotted are probably the most useful)}

\item{arrow_colour}{Colour for the process step arrow}

\item{arrow_style}{Arrow style for use by DiagrrammeR::create_edge_df
(tapered and solid are probably the most useful)}

\item{colour_palette}{Colour palette to use for multiple lanes}

\item{lane_colour}{Colour for dividing line between lanes}

\item{title_fontsize}{Size of the title font in points (NA means no title)}

\item{fontname}{Name of font to use for process steps and arrows}

\item{title_fontname}{Name of font for title}

\item{ctoffset}{How much offset to use for cycle time arrows}
}
\value{
graph Graph that can be used by DiagrammeR for render or export
}
\description{
Create a graph in the form of a swimlane map for the process
}
\examples{
# Create a template process with 5 steps and fill out a standard template for the SIPOC
# Note that row 1 contains L1 (level 1) information and the subsequent 4 rows contain L2
# steps that will would appear in a SIPOC
p <- create_process(steps = 5, sipoc_template = TRUE)

# You could then edit the process in R or export to Excel then re-import
# graph_slm plots all the steps in the process. If an R column is included, that changes
# the name of the swinlane
# need to specify the parent step number
g <- graph_slm(p)

# Visualize the process as a graph using Diagrammer
DiagrammeR::render_graph(g)
}
