% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_tree.R
\name{graph_tree}
\alias{graph_tree}
\title{Create a tree graph from a process}
\usage{
graph_tree(
  process,
  depth = NA,
  numbering = FALSE,
  loopback = FALSE,
  resource = FALSE,
  step_colour = base_colour_palette,
  font_colour = base_text_palette,
  box_colour = base_box_palette,
  arrow_colour = base_arrow_palette,
  step_textwrap = 18,
  step_width = 1.25,
  step_height = 0.5,
  fontname = "Calibri,Arial,Helvetica",
  orientation = "LR",
  xsep = 0.25,
  ysep = 0.25
)
}
\arguments{
\item{process}{A Process data frame or tibble}

\item{depth}{The deepest level to plot}

\item{numbering}{Number the process steps}

\item{loopback}{whether to draw an edge from last sub-process}

\item{resource}{whether to show resources on the graph}

\item{step_colour}{set of colours to fill process steps}

\item{font_colour}{Set of colours for process step text}

\item{box_colour}{set of colours for edges of each process step and to show resources}

\item{arrow_colour}{Colour for the process step arrow}

\item{step_textwrap}{String length for wrapping text in process boxes (nodes)}

\item{step_width}{Minimum width for process step box}

\item{step_height}{Minimum height for process step box}

\item{fontname}{Font for text in the graph}

\item{orientation}{How to layout the tree LR (left-to-right), RL, TB (top-to-bottom) or BT}

\item{xsep}{Minimum separation between nodes horizontally}

\item{ysep}{Minimum separation between nodes vertically}
}
\value{
A graph object of class dgr_graph
}
\description{
This function takes a process object and outputs the graph for the process as a tree
}
