% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_edges.R
\name{label_edges}
\alias{label_edges}
\title{Creates edges from a process with labels}
\usage{
label_edges(process, terminals = FALSE)
}
\arguments{
\item{process}{Process dataframe in the standard format}

\item{terminals}{specify whether to create initial output and final output edges}
}
\value{
pairs A dataframe with columns from, to and label indicating the edges of the graph
}
\description{
Uses process dataframe including inputs and outputs to each step to create a
dataframe that can be used by DiagrammeR to create a graph. It is normally used within
process_plot_Lx to create labels at one specific process label
}
