% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_subprocesses.R
\name{list_subprocesses}
\alias{list_subprocesses}
\title{Get all subprocesses for a process}
\usage{
list_subprocesses(process, depth = NA)
}
\arguments{
\item{process}{Input process}

\item{depth}{maximum depth}
}
\value{
subprocesses datafrane with matching Level and Parent
}
\description{
Get a list of all subprocesses as dataframe
}
\examples{
p <- create_process(steps = 5, sipoc_template = TRUE)
list_subprocesses(p)

}
