% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_steps.R
\name{number_steps}
\alias{number_steps}
\title{process_label}
\usage{
number_steps(process)
}
\arguments{
\item{proces}{process in a standard format}
}
\value{
a process object with additional columns IDdotted and one LxID column for
each Lx column
}
\description{
Create individual step numbers (eg 1.1, 1.2.3) at all levels
return a process with additional column IDdotted
}
\examples{
process <- number_steps(process)

}
