% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_utils.R
\name{process_L2_title}
\alias{process_L2_title}
\title{process_L2_title}
\usage{
process_L2_title(process, L2ID)
}
\arguments{
\item{process}{process in a standard format}

\item{L2ID}{the ID of the L2 process of interest}
}
\value{
Name of the L2 process as a character string
}
\description{
Find the name of an L2 process
}
\examples{
\dontrun{
title <- process_L2_title(process, L2ID)
}
}
