% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_utils.R
\name{process_L2IDs}
\alias{process_L2IDs}
\title{process_L2IDs}
\usage{
process_L2IDs(process)
}
\arguments{
\item{process}{process in a standard format}
}
\value{
Array of L2 IDs for a process
}
\description{
Find all the L2 IDs for a process
}
\examples{
\dontrun{
title <- process_L2IDs(process)
}
}
