% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_utils.R
\name{process_title}
\alias{process_title}
\title{process_title}
\usage{
process_title(process)
}
\arguments{
\item{process}{process in a standard format}
}
\value{
Name of the process as a character string
}
\description{
Find the name of the process as described in L1
}
\examples{
\dontrun{
title <- process_title(process)
}

}
