% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_subprocess.R
\name{select_subprocess}
\alias{select_subprocess}
\title{Select a subprocess}
\usage{
select_subprocess(process, level = 1, parent = "", depth = NA)
}
\arguments{
\item{process}{Complete process to extract subprocess from}

\item{level}{Top level to include}

\item{parent}{Parent of selected level}

\item{max_level}{Deepest level to include NA means no limit}
}
\description{
Extract a coherent subprocess from a complete process
}
