% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{text_palette}
\alias{text_palette}
\title{text_palette}
\usage{
text_palette(n = 5)
}
\arguments{
\item{n}{number of text colours requested}
}
\value{
colours vector of colours in hex-format
}
\description{
Return a colour palette for text suitable for process step colours
}
