% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_process.R
\name{validate_process}
\alias{validate_process}
\title{Check process tibble for essential elements}
\usage{
validate_process(process, allow_empty = FALSE)
}
\arguments{
\item{process}{Process tibble to check}

\item{allow_empty}{Allow an empty proces to validate}
}
\value{
status Logical
}
\description{
This function helps when loading a process to ensure it has the correct columns
}
\examples{
p <- create_process(step = 5, sipoc_template = TRUE)
if (validate_process(p)) print("Process Valid")
}
