% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_process.R
\name{wrap_process}
\alias{wrap_process}
\title{Return coordinates for a process}
\usage{
wrap_process(
  process,
  terminals = FALSE,
  xscale = 2,
  yscale = 1.5,
  wrap = 5,
  branchoffset = 0.5,
  tscale = 1,
  ctoffset = 0.3
)
}
\arguments{
\item{process}{Process in standard format}

\item{terminals}{Include initial input and final output}

\item{xscale}{Scale factor for x-dimension}

\item{yscale}{Scale factor for y-dimension}

\item{wrap}{Maximum number of steps before wrapping}

\item{branchoffset}{How much of scale to take up for multiple branches}

\item{tscale}{Relative location of terminal nodes}

\item{ctoffset}{Relative location of the cycle time double arrow in vertical direction}
}
\description{
This function takes a process - which include branching - and returns xy coordinates as a
data frame which can be passed to DiagrammeR functions to specify node coordinates
}
